Write inst device flow to tide file
This commit adds codes to write instantaneous device flow to the tide
file. The dataset name is 'inst device flow', and the dimension is (#
gates, max_dev = 10).
No gate and device information is written yet.

DSM2-1222

+++ dsm2/src/hdf_tidefile/hdf5_write_attributes.f90
@@ -76,7 +76,7 @@ subroutine hdf5_write_attributes()
 
     integer, dimension(1) :: hdf5_dummy_integer
     integer(SIZE_T) :: hdf5_int_size
-    integer :: scalar
+    integer(SIZE_T) :: scalar
     real(gtm_real), dimension(1) :: hdf5_dummy_real
 
     integer cdt2jmin
